tests/wpt/web-platform-tests/css/css-text-decor/parsing/text-decoration-style-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Decoration Test: getComputedStyle().textDecorationStyle</title>
<link rel="help" href="https://drafts.csswg.org/css-text-decor-3/#text-decoration-style-property">
<meta name="assert" content="text-decoration-style computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("text-decoration-style", "solid");
test_computed_value("text-decoration-style", "double");
test_computed_value("text-decoration-style", "dotted");
test_computed_value("text-decoration-style", "dashed");
test_computed_value("text-decoration-style", "wavy");
</script>
</body>
</html>
